@@ -6,6 +6,7 @@
 
 import 'dart:convert';
 
+import 'package:fair/src/internal/error_tips.dart';
 import 'package:fair/src/runtime/fair_message_dispatcher.dart';
 import 'package:fair/src/runtime/runtime_fair_delegate.dart';
 import 'package:flutter/foundation.dart';
@@ -98,6 +99,11 @@ class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientM
   FairApp? _fairApp;
   String? bundleType;
   late String state2key;
+  bool isLoadJsError = false;
+  String? loadJsErrorInfo;
+  String? jsSource;
+  String? rawJsPath;
+  int? loadJsErrorLineNumber;
 
   // None nullable
   late FairDelegate delegate;
@@ -128,7 +134,36 @@ class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientM
     }
     // if it's not in a tree, it's not unnecessary to load js any more.
     if(mounted) {
-      await Future.wait([_mFairApp.runtime.addScript(state2key, resolveJS, widget.data), _mFairApp.register(this)]);
+      final results = await Future.wait([
+        _mFairApp.runtime.addScript(state2key, resolveJS, widget.data),
+        _mFairApp.register(this)
+      ]);
+
+      //debug mode throw error message to WarningWidget
+      if (!kReleaseMode && results.isNotEmpty) {
+        print('addScript Result:${results.first}');
+        try {
+          final addScriptResult = results.first.toString();
+          var errorResult = jsonDecode(addScriptResult);
+          isLoadJsError = (errorResult['status'] == 'error');
+
+          if (isLoadJsError) {
+            loadJsErrorInfo = errorResult['errorInfo'] ?? '';
+            loadJsErrorLineNumber = errorResult['lineNumber'];
+
+            //debug mode use json
+            if(widget.path?.endsWith('.fair.json') == true){
+              rawJsPath = widget.path?.replaceFirst('.fair.json', '.fair.js') ??
+                  widget.path;
+
+              jsSource = await rootBundle.loadString(rawJsPath!);
+            }
+          }
+        } catch (e) {
+          print(e);
+        }
+      }
+
       if(mounted) {
         delegate.didChangeDependencies();
         _reload();
@@ -146,10 +181,13 @@ class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientM
     if (mounted) {
       var name = state2key;
       var path = widget.path ?? _fairApp!.pathOfBundle(widget.name ?? '');
+      if (path?.endsWith('.js') == true) {
+        path = path?.replaceFirst(RegExp(r"\.(js)$"), ".bin");
+      }
       bundleType =
-          widget.path != null && widget.path?.startsWith('http') == true
-              ? 'Http'
-              : 'Asset';
+      widget.path != null && widget.path?.startsWith('http') == true
+          ? 'Http'
+          : 'Asset';
 
       parse(context, page: name, url: path, data: widget.data ?? {})
           .then((value) {
@@ -164,16 +202,30 @@ class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientM
   Widget build(BuildContext context) {
     super.build(context);
     assert(_fairApp != null, 'FairWidget must be descendant of FairApp');
+    if (!kReleaseMode && isLoadJsError) {
+      return Scaffold(
+        body: WarningWidget(
+          parentContext: context,
+          url: rawJsPath,
+          error: loadJsErrorInfo,
+          stackTrace: jsSource,
+          highlightLines: [loadJsErrorLineNumber ?? -1],
+          solution:
+          'Unsupported JavaScript syntax, please check and correct your Dart method coding!',
+        ),
+      );
+    }
     var builder = widget.holder ?? _fairApp?.placeholderBuilder;
     var result = _child ?? builder?.call(context);
     if (!kReleaseMode && _fairApp!.debugShowFairBanner) {
-      result = _CheckedModeBanner(bundleType??'', child: result??Container());
+      result =
+          _CheckedModeBanner(bundleType ?? '', child: result ?? Container());
     }
-    return result??Container();
+    return result ?? Container();
   }
 
   @override
-  bool get wantKeepAlive => widget.wantKeepAlive??false;
+  bool get wantKeepAlive => widget.wantKeepAlive ?? false;
 
   /// Implementations of this method should end with a call to the inherited
   /// method, as in `super.dispose()`.
@@ -189,9 +241,9 @@ class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientM
 
   @override
   void call(String t) {
-    var params={};
+    var params = {};
     try {
-      params= jsonDecode(t);
+      params = jsonDecode(t);
     } catch (e) {
       print(e);
     }
@@ -205,7 +257,7 @@ class FairState extends State<FairWidget> with Loader, AutomaticKeepAliveClientM
 /// Delegate for business logic. The delegate share similar life-circle with [State].
 class FairDelegate extends RuntimeFairDelegate {
   FairState? _state;
- late String _key;
+  late String _key;
 
   void _bindState(FairState? state) {
     assert(state != null, 'FairState should not be null');
