@@ -1,1 +1,1 @@
-GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function MomentsModel(){const inner=MomentsModel.__inner__;if(this==__global__){return new MomentsModel({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);MomentsModel.prototype.ctor.apply(this,args);return this;}}MomentsModel.__inner__=function inner(){this.avatar='';this.username='';this.content='';this.picture='';this.likes='';this.comments='';};MomentsModel.prototype={};MomentsModel.prototype.ctor=function(){};function _MomentsListState(){const inner=_MomentsListState.__inner__;if(this==__global__){return new _MomentsListState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_MomentsListState.prototype.ctor.apply(this,args);return this;}}_MomentsListState.__inner__=function inner(){this._listData=[];this._page=0;};_MomentsListState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairCommonPlugin().http(convertObjectLiteralToSetOrMap({['pageName']:'#FairKey#',['method']:'GET',['url']:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/ed04fcaf655b79b1ae81ad2836ce67e9_moments_data.json',['data']:convertObjectLiteralToSetOrMap({['page']:_page,}),['callback']:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=MomentsModel();try{dataItem.avatar=item.avatar;dataItem.username=item.username;dataItem.content=item.content;dataItem.picture=item.picture;dataItem.likes=item.likes;dataItem.comments=item.comments;}catch(e){dataItem.avatar=item.__op_idx__('avatar');dataItem.username=item.__op_idx__('username');dataItem.content=item.__op_idx__('content');dataItem.picture=item.__op_idx__('picture');dataItem.likes=item.__op_idx__('likes');dataItem.comments=item.__op_idx__('comments');}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});},}));}},dataLength:function dataLength(){const __thiz__=this;with(__thiz__){return _listData.length;}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},_getAvatar:function _getAvatar(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).avatar;}}},_getUsername:function _getUsername(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).username;}}},_getContent:function _getContent(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).content;}}},_getPicture:function _getPicture(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).picture;}}},_getLikes:function _getLikes(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).likes;}}},_getComments:function _getComments(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).comments;}}},};_MomentsListState.prototype.ctor=function(){};;return _MomentsListState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
+GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function MomentsModel(){const inner=MomentsModel.__inner__;if(this==__global__){return new MomentsModel({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);MomentsModel.prototype.ctor.apply(this,args);return this;}}MomentsModel.__inner__=function inner(){this.avatar='';this.username='';this.content='';this.picture='';this.likes='';this.comments='';};MomentsModel.prototype={};MomentsModel.prototype.ctor=function(){};function _MomentsListState(){const inner=_MomentsListState.__inner__;if(this==__global__){return new _MomentsListState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_MomentsListState.prototype.ctor.apply(this,args);return this;}}_MomentsListState.__inner__=function inner(){this._listData=[];this._page=0;};_MomentsListState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairNet.requestData({method:FairNet.GET,url:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/ed04fcaf655b79b1ae81ad2836ce67e9_moments_data.json',data:convertObjectLiteralToSetOrMap({['page']:_page,}),success:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=MomentsModel();try{dataItem.avatar=item.__op_idx__('avatar');dataItem.username=item.__op_idx__('username');dataItem.content=item.__op_idx__('content');dataItem.picture=item.__op_idx__('picture');dataItem.likes=item.__op_idx__('likes');dataItem.comments=item.__op_idx__('comments');}catch(e){}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});}});}},dataLength:function dataLength(){const __thiz__=this;with(__thiz__){return _listData.length;}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},_getAvatar:function _getAvatar(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).avatar;}}},_getUsername:function _getUsername(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).username;}}},_getContent:function _getContent(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).content;}}},_getPicture:function _getPicture(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).picture;}}},_getLikes:function _getLikes(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).likes;}}},_getComments:function _getComments(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).comments;}}},};_MomentsListState.prototype.ctor=function(){};;return _MomentsListState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
