@@ -1,1 +1,1 @@
-GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function ItemData(){const inner=ItemData.__inner__;if(this==__global__){return new ItemData({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);ItemData.prototype.ctor.apply(this,args);return this;}}ItemData.__inner__=function inner(){this.icon='';this.title='';this.subTitle='';this.distance='';};ItemData.prototype={};ItemData.prototype.ctor=function(){};function _ListDemoPageState(){const inner=_ListDemoPageState.__inner__;if(this==__global__){return new _ListDemoPageState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_ListDemoPageState.prototype.ctor.apply(this,args);return this;}}_ListDemoPageState.__inner__=function inner(){this._listData=[];this._page=0;};_ListDemoPageState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairCommonPlugin().http(convertObjectLiteralToSetOrMap({['pageName']:'#FairKey#',['method']:'GET',['url']:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/3b8ae7a4e0884b4d75b8094f6c83cd8c_list_page_data.json',['data']:convertObjectLiteralToSetOrMap({['page']:_page,}),['callback']:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=ItemData();try{dataItem.icon=item.icon;dataItem.title=item.title;dataItem.subTitle=item.subTitle;dataItem.distance=item.distance;}catch(e){dataItem.icon=item.__op_idx__('icon');dataItem.title=item.__op_idx__('title');dataItem.subTitle=item.__op_idx__('subTitle');dataItem.distance=item.__op_idx__('distance');}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});},}));}},dataLength:function dataLength(){const __thiz__=this;with(__thiz__){return _listData.length;}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},_getIcon:function _getIcon(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).icon;}}},_getTitle:function _getTitle(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).title;}}},_getSubTitle:function _getSubTitle(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).subTitle;}}},_getDistance:function _getDistance(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).distance;}}},};_ListDemoPageState.prototype.ctor=function(){};;return _ListDemoPageState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
+GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function ItemData(){const inner=ItemData.__inner__;if(this==__global__){return new ItemData({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);ItemData.prototype.ctor.apply(this,args);return this;}}ItemData.__inner__=function inner(){this.icon='';this.title='';this.subTitle='';this.distance='';};ItemData.prototype={};ItemData.prototype.ctor=function(){};function _ListDemoPageState(){const inner=_ListDemoPageState.__inner__;if(this==__global__){return new _ListDemoPageState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_ListDemoPageState.prototype.ctor.apply(this,args);return this;}}_ListDemoPageState.__inner__=function inner(){this._listData=[];this._page=0;};_ListDemoPageState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairNet.requestData({method:FairNet.GET,url:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/3b8ae7a4e0884b4d75b8094f6c83cd8c_list_page_data.json',data:convertObjectLiteralToSetOrMap({['page']:_page,}),success:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=ItemData();try{dataItem.icon=item.__op_idx__('icon');dataItem.title=item.__op_idx__('title');dataItem.subTitle=item.__op_idx__('subTitle');dataItem.distance=item.__op_idx__('distance');}catch(e){}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});}});}},dataLength:function dataLength(){const __thiz__=this;with(__thiz__){return _listData.length;}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},_getIcon:function _getIcon(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).icon;}}},_getTitle:function _getTitle(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).title;}}},_getSubTitle:function _getSubTitle(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).subTitle;}}},_getDistance:function _getDistance(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).distance;}}},};_ListDemoPageState.prototype.ctor=function(){};;return _ListDemoPageState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
