@@ -1,1 +1,1 @@
-GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function StaggeredViewModel(){const inner=StaggeredViewModel.__inner__;if(this==__global__){return new StaggeredViewModel({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);StaggeredViewModel.prototype.ctor.apply(this,args);return this;}}StaggeredViewModel.__inner__=function inner(){this.picUrl='';this.ratio=1.5;};StaggeredViewModel.prototype={};StaggeredViewModel.prototype.ctor=function(){};function _StaggeredItemViewState(){const inner=_StaggeredItemViewState.__inner__;if(this==__global__){return new _StaggeredItemViewState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_StaggeredItemViewState.prototype.ctor.apply(this,args);return this;}}_StaggeredItemViewState.__inner__=function inner(){this._listData=[];this._page=0;};_StaggeredItemViewState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairCommonPlugin().http(convertObjectLiteralToSetOrMap({['pageName']:'#FairKey#',['method']:'GET',['url']:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/3a0809703d09800b54491c4d1fc710da_staggeredview_data.json',['data']:convertObjectLiteralToSetOrMap({['page']:_page,}),['callback']:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=StaggeredViewModel();try{dataItem.picUrl=item.imagePath;dataItem.ratio=item.ratio;}catch(e){dataItem.picUrl=item.__op_idx__('imagePath');dataItem.ratio=item.__op_idx__('ratio');}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});},}));}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},};_StaggeredItemViewState.prototype.ctor=function(){};;return _StaggeredItemViewState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
+GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function StaggeredViewModel(){const inner=StaggeredViewModel.__inner__;if(this==__global__){return new StaggeredViewModel({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);StaggeredViewModel.prototype.ctor.apply(this,args);return this;}}StaggeredViewModel.__inner__=function inner(){this.picUrl='';this.ratio=1.5;};StaggeredViewModel.prototype={};StaggeredViewModel.prototype.ctor=function(){};function _StaggeredItemViewState(){const inner=_StaggeredItemViewState.__inner__;if(this==__global__){return new _StaggeredItemViewState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_StaggeredItemViewState.prototype.ctor.apply(this,args);return this;}}_StaggeredItemViewState.__inner__=function inner(){this._listData=[];this._page=0;};_StaggeredItemViewState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairNet.requestData({method:FairNet.GET,url:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/3a0809703d09800b54491c4d1fc710da_staggeredview_data.json',data:convertObjectLiteralToSetOrMap({['page']:_page,}),success:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=StaggeredViewModel();try{dataItem.picUrl=item.__op_idx__('imagePath');dataItem.ratio=item.__op_idx__('ratio');}catch(e){}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});}});}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},};_StaggeredItemViewState.prototype.ctor=function(){};;return _StaggeredItemViewState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
