@@ -1,1 +1,1 @@
-GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function TabBarModel(){const inner=TabBarModel.__inner__;if(this==__global__){return new TabBarModel({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);TabBarModel.prototype.ctor.apply(this,args);return this;}}TabBarModel.__inner__=function inner(){this.imagePath='';this.title='';};TabBarModel.prototype={};TabBarModel.prototype.ctor=function(){};function _SugarTabBarPageState(){const inner=_SugarTabBarPageState.__inner__;if(this==__global__){return new _SugarTabBarPageState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_SugarTabBarPageState.prototype.ctor.apply(this,args);return this;}}_SugarTabBarPageState.__inner__=function inner(){this._listData=[];this._page=0;};_SugarTabBarPageState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairCommonPlugin().http(convertObjectLiteralToSetOrMap({['pageName']:'#FairKey#',['method']:'GET',['url']:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/81900c8ba18b20328389b43c729a251e_tabbar_data.json',['data']:convertObjectLiteralToSetOrMap({['page']:_page,}),['callback']:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=TabBarModel();try{dataItem.imagePath=item.imageUrl;dataItem.title=item.title;}catch(e){dataItem.imagePath=item.__op_idx__('imageUrl');dataItem.title=item.__op_idx__('title');}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});},}));}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},_getImagePath:function _getImagePath(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).imagePath;}}},_getTitle:function _getTitle(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).title;}}},};_SugarTabBarPageState.prototype.ctor=function(){};;return _SugarTabBarPageState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
+GLOBAL['#FairKey#']=(function(__initProps__){const __global__=this;return runCallback(function(__mod__){with(__mod__.imports){function TabBarModel(){const inner=TabBarModel.__inner__;if(this==__global__){return new TabBarModel({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);TabBarModel.prototype.ctor.apply(this,args);return this;}}TabBarModel.__inner__=function inner(){this.imagePath='';this.title='';};TabBarModel.prototype={};TabBarModel.prototype.ctor=function(){};function _SugarTabBarPageState(){const inner=_SugarTabBarPageState.__inner__;if(this==__global__){return new _SugarTabBarPageState({__args__:arguments});}else{const args=arguments.length>0?arguments[0].__args__||arguments:[];inner.apply(this,args);_SugarTabBarPageState.prototype.ctor.apply(this,args);return this;}}_SugarTabBarPageState.__inner__=function inner(){this._listData=[];this._page=0;};_SugarTabBarPageState.prototype={onLoad:function onLoad(){const __thiz__=this;with(__thiz__){requestData();}},onUnload:function onUnload(){const __thiz__=this;with(__thiz__){}},requestData:function requestData(){const __thiz__=this;with(__thiz__){_page++;FairNet.requestData({method:FairNet.GET,url:'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/81900c8ba18b20328389b43c729a251e_tabbar_data.json',data:convertObjectLiteralToSetOrMap({['page']:_page,}),success:function dummy(resp){if(resp==null){return null;}let data=resp.__op_idx__('data');data.forEach(function dummy(item){let dataItem=TabBarModel();try{dataItem.imagePath=item.__op_idx__('imageUrl');dataItem.title=item.__op_idx__('title');}catch(e){}_listData.add(dataItem);});setState('#FairKey#',function dummy(){});}});}},isDataEmpty:function isDataEmpty(){const __thiz__=this;with(__thiz__){return _listData.isEmpty;}},_getImagePath:function _getImagePath(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).imagePath;}}},_getTitle:function _getTitle(index){const __thiz__=this;const __arg_ctx__={index,};with(__thiz__){with(__arg_ctx__){return _listData.__op_idx__(index).title;}}},};_SugarTabBarPageState.prototype.ctor=function(){};;return _SugarTabBarPageState();}},[]);})(convertObjectLiteralToSetOrMap(JSON.parse('#FairProps#')));
