@@ -3,8 +3,8 @@ import 'dart:math';
 import 'package:example/fair_widget/delegate/test_fair_delegate.dart';
 import 'package:example/fair_widget/plugin/fair_basic_plugin.dart';
 import 'package:example/home_page.dart';
-import 'package:example/plugins/fair_common_plugin.dart';
 import 'package:fair/fair.dart';
+import 'package:fair_extension/fair_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -23,23 +23,21 @@ void main() async {
   ]);
 
   FairApp.runApplication(
-    FairApp(
-      child: MyApp(),
-      delegate: {
-        ///此处delegate注册的key名必须与fairwidget页面name的名字一致,
-        ///TestFairDelegate只作用于相同名字的fairwidget
-        'assets/fair/lib_fair_widget_fair_delegate_widget.fair.json':
-            (ctx, _) => TestFairDelegate(),
-      },
-      generated: g.FairAppModule(),
-    ),
+      FairApp(
+        child: MyApp(),
+        delegate: {
+          ///此处delegate注册的key名必须与fairwidget页面name的名字一致,
+          ///TestFairDelegate只作用于相同名字的fairwidget
+          'assets/fair/lib_fair_widget_fair_delegate_widget.fair.json':
+              (ctx, _) => TestFairDelegate(),
+        },
+        generated: g.FairAppModule(),
+      ),
 
-    ///需要在此注册需要全局使用的plugin,key名可以随意不做要求
-    plugins: {
-      'FairBasicPlugin': FairBasicPlugin(),
-      'FairCommonPlugin': FairCommonPlugin()
-    },
-  );
+      ///需要在此注册需要全局使用的plugin,key名可以随意不做要求
+      plugins: FairExtension.plugins
+        ..addAll({'FairBasicPlugin': FairBasicPlugin()}),
+      jsPlugins: FairExtension.jsPlugins);
 }
 
 class MyApp extends StatelessWidget {
@@ -50,7 +48,8 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        home: HomePage());
+        home: HomePage(),
+        navigatorKey: FairExtension.fairNavigatorKey);
   }
 }
 
