@@ -1,5 +1,8 @@
 import 'package:example/plugins/fair_common_plugin.dart';
 import 'package:fair/fair.dart';
+import 'package:fair_extension/log/log.dart';
+import 'package:fair_extension/navigator/fair_navigator_plugin.dart';
+import 'package:fair_extension/net/fair_net_plugin.dart';
 import 'package:flutter/material.dart';
 
 @FairPatch()
@@ -24,29 +27,25 @@ class _HomeScrollViewState extends State<HomeScrollView> {
 
   void requestData() {
     _page++;
-    FairCommonPlugin().http({
-      'pageName': '#FairKey#',
-      'method': 'GET',
-      'url':
-          'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/3158c2fc5e3ed9bc08b34f8d694c763d_home_scroll_data.json',
-      'data': {'page': _page},
-      'callback': (resp) {
-        if (resp == null) {
-          return;
-        }
-        var data = resp['data'];
-        data.forEach((item) {
-          var dataItem = HomeItemData();
-          try {
-            dataItem.imagePath = item.imageUrl;
-          } catch (e) {
-            dataItem.imagePath = item['imageUrl'];
+    FairNet.requestData(
+        method: FairNet.GET,
+        url:
+            'https://wos2.58cdn.com.cn/DeFazYxWvDti/frsupload/3158c2fc5e3ed9bc08b34f8d694c763d_home_scroll_data.json',
+        data: {'page': _page},
+        success: (resp) {
+          if (resp == null) {
+            return;
           }
-          _listData.add(dataItem);
+          var data = resp['data'];
+          data.forEach((item) {
+            var dataItem = HomeItemData();
+            try {
+              dataItem.imagePath = item['imageUrl'];
+            } catch (e) {}
+            _listData.add(dataItem);
+          });
+          setState(() {});
         });
-        setState(() {});
-      }
-    });
   }
 
   bool isDataEmpty() {
@@ -93,19 +92,22 @@ class _HomeScrollViewState extends State<HomeScrollView> {
               crossAxisSpacing: 10.0,
               childAspectRatio: 2.0,
             ),
-            delegate: SliverChildBuilderDelegate(
-              Sugar.nullableIndexedWidgetBuilder(
-                (context, index) => AspectRatio(
-                  aspectRatio: 1.5,
-                  child: ClipRRect(
-                    borderRadius: const BorderRadius.all(Radius.circular(4.0)),
-                    child:
-                        Image.network(_getImagePath(index), fit: BoxFit.cover),
-                  ),
-                ),
-              ),
-              childCount: 4,
-            ),
+            delegate: Sugar.sliverChildBuilderDelegate(
+                builder: (content, index) {
+                  return GestureDetector(
+                    onTap: _onTapIndex(index),
+                    child: AspectRatio(
+                      aspectRatio: 1.5,
+                      child: ClipRRect(
+                        borderRadius:
+                            const BorderRadius.all(Radius.circular(4.0)),
+                        child: Image.network(_getImagePath(index),
+                            fit: BoxFit.cover),
+                      ),
+                    ),
+                  );
+                },
+                childCount: 4),
           ),
 
           ///列表
@@ -127,6 +129,14 @@ class _HomeScrollViewState extends State<HomeScrollView> {
         ],
       ),
     ));
+  }
+
+  _onTapIndex(int index) {
+    FairLog.log('点击index >>> $index');
+    FairNavigator.pushFairPath(
+        fairPath:
+            'assets/fair/lib_template_hotel_listview_hotel_listview_template.fair.json',
+        fairName: 'lib_template_hotel_listview_hotel_listview_template');
   }
 }
 
