@@ -34,8 +34,7 @@ class _State extends State<IfEqualPage> {
 
   void onTapText() {
     _count = _count + 1;
-    setState(() {
-    });
+    setState(() {});
   }
 
   @override
@@ -54,9 +53,9 @@ class _State extends State<IfEqualPage> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            Sugar.ifEqual(_count, 2,
-                falseValue: ()=>Image.asset('assets/image/logo.png'),
-                trueValue: ()=>Image.asset('assets/image/logo2.png')),
+            // Sugar.ifEqual(_count, 2,
+            //     falseValue: Image.asset('assets/image/logo.png'),
+            //     trueValue: Image.asset('assets/image/logo2.png')),
             Padding(
               padding: EdgeInsets.only(top: 20),
               child: Text('_count = $_count'),
