@@ -61,9 +61,9 @@ class _State extends State<IfEqualBoolPage> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            Sugar.ifEqualBool(_countCanMod2(),
-                falseValue: ()=>Image.asset('assets/image/logo.png'),
-                trueValue: ()=>Image.asset('assets/image/logo2.png')),
+            // Sugar.ifEqualBool(_countCanMod2(),
+            //     falseValue: Image.asset('assets/image/logo.png'),
+            //     trueValue: Image.asset('assets/image/logo2.png')),
             Padding(
               padding: EdgeInsets.only(top: 20),
               child: Text('_count = $_count'),
